+++ mithril-common/src/messages/snapshot_list.rs
@@ -21,14 +21,22 @@ pub struct SnapshotListItemMessage {
     /// Hash of the associated certificate
     pub certificate_hash: String,
 
-    /// Size of the snapshot file in Bytes
+    /// Size of the immutables snapshot file in Bytes
     pub size: u64,
+
+    /// Size of the ancillary files in Bytes
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub ancillary_size: Option<u64>,
 
     /// Date and time at which the snapshot was created
     pub created_at: DateTime<Utc>,
 
-    /// Locations where the binary content of the snapshot can be retrieved
+    /// Locations where the snapshot of the immutable files can be retrieved
     pub locations: Vec<String>,
+
+    /// Locations where the snapshot of the ancillary files can be retrieved
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub ancillary_locations: Option<Vec<String>>,
 
     /// Compression algorithm of the snapshot archive
     pub compression_algorithm: CompressionAlgorithm,
@@ -50,10 +58,12 @@ impl SnapshotListItemMessage {
             certificate_hash: "d5daf6c03ace4a9c074e951844075b9b373bafc4e039160e3e2af01823e9abfb"
                 .to_string(),
             size: 807803196,
+            ancillary_size: Some(123456789),
             created_at: DateTime::parse_from_rfc3339("2023-01-19T13:43:05.618857482Z")
                 .unwrap()
                 .with_timezone(&Utc),
             locations: vec!["https://host/certificate.tar.gz".to_string()],
+            ancillary_locations: Some(vec!["https://host/ancillary.tar.gz".to_string()]),
             compression_algorithm: CompressionAlgorithm::default(),
             cardano_node_version: "0.0.1".to_string(),
         }
@@ -64,6 +74,21 @@ impl SnapshotListItemMessage {
 mod tests {
     use super::*;
 
+    pub type SnapshotListMessageUntilV0_1_47 = Vec<SnapshotListItemMessageUntilV0_1_47>;
+
+    #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+    pub struct SnapshotListItemMessageUntilV0_1_47 {
+        pub digest: String,
+        pub network: String,
+        pub beacon: CardanoDbBeacon,
+        pub certificate_hash: String,
+        pub size: u64,
+        pub created_at: DateTime<Utc>,
+        pub locations: Vec<String>,
+        pub compression_algorithm: CompressionAlgorithm,
+        pub cardano_node_version: String,
+    }
+
     const CURRENT_JSON: &str = r#"[{
         "digest": "0b9f5ad7f33cc523775c82249294eb8a1541d54f08eb3107cafc5638403ec7c6",
         "network": "preview",
@@ -73,16 +98,20 @@ mod tests {
         },
         "certificate_hash": "d5daf6c03ace4a9c074e951844075b9b373bafc4e039160e3e2af01823e9abfb",
         "size": 807803196,
+        "ancillary_size": 123456789,
         "created_at": "2023-01-19T13:43:05.618857482Z",
         "locations": [
             "https://host/certificate.tar.gz"
+        ],
+        "ancillary_locations": [
+            "https://host/ancillary.tar.gz"
         ],
         "compression_algorithm": "zstandard",
         "cardano_node_version": "1.0.0"
     }]"#;
 
-    fn golden_message_current() -> SnapshotListMessage {
-        vec![SnapshotListItemMessage {
+    fn golden_message_until_open_api_0_1_47() -> SnapshotListMessageUntilV0_1_47 {
+        vec![SnapshotListItemMessageUntilV0_1_47 {
             digest: "0b9f5ad7f33cc523775c82249294eb8a1541d54f08eb3107cafc5638403ec7c6".to_string(),
             network: "preview".to_string(),
             beacon: CardanoDbBeacon {
@@ -99,6 +128,36 @@ mod tests {
             compression_algorithm: CompressionAlgorithm::Zstandard,
             cardano_node_version: "1.0.0".to_string(),
         }]
+    }
+
+    fn golden_message_current() -> SnapshotListMessage {
+        vec![SnapshotListItemMessage {
+            digest: "0b9f5ad7f33cc523775c82249294eb8a1541d54f08eb3107cafc5638403ec7c6".to_string(),
+            network: "preview".to_string(),
+            beacon: CardanoDbBeacon {
+                epoch: Epoch(86),
+                immutable_file_number: 1728,
+            },
+            certificate_hash: "d5daf6c03ace4a9c074e951844075b9b373bafc4e039160e3e2af01823e9abfb"
+                .to_string(),
+            size: 807803196,
+            ancillary_size: Some(123456789),
+            created_at: DateTime::parse_from_rfc3339("2023-01-19T13:43:05.618857482Z")
+                .unwrap()
+                .with_timezone(&Utc),
+            locations: vec!["https://host/certificate.tar.gz".to_string()],
+            ancillary_locations: Some(vec!["https://host/ancillary.tar.gz".to_string()]),
+            compression_algorithm: CompressionAlgorithm::Zstandard,
+            cardano_node_version: "1.0.0".to_string(),
+        }]
+    }
+
+    #[test]
+    fn test_current_json_deserialized_into_message_supported_until_open_api_0_1_47() {
+        let json = CURRENT_JSON;
+        let message: SnapshotListMessageUntilV0_1_47 = serde_json::from_str(json).unwrap();
+
+        assert_eq!(golden_message_until_open_api_0_1_47(), message);
     }
 
     #[test]